configure.in: 	print config info for sound type source/quakefs.c: 	fix fs_userpath handling for other game types and add support for 	abyss source/qargs.c: 	add -abyss checking include/qdefs.h: 	add abyss support
Rudimentary support for the abyss mission pack has been added. Currently this
involves just the addition of -abyss to the supported command line options
(acts just like -rogue and -hipnotic) and adding abyss to the list of game
directories. Any other support that's needed (as with other rogue and hipnotic
modifications) has not been added as it is not known what's needed.