fix the 0 byte udp packet DoS bug (for linux, anyway: mods from ProQuake applied for windows, but I can't test).
In, linux, the problem was caused by nuq checking to see if there were any
bytes to be read and if there were 0 bytes, just carying on without reading
anything. It seems that the 0 byte packet was blocking the queue and so a 0
byte recvfrom is needed to make sure it gets removed.